Updated genSurvey and better loading

+++ genSurvey.m
@@ -1,8 +1,10 @@
-clearvars
+function genSurvey(data, filename)
 
 % Delete and remake file
-filename = 'MatlabTest.txt';
-system(['rm ' filename]);
+% filename = '10Trial2.txt';
+if exist(filename, 'file')
+    system(['rm ' filename]);
+end
 system(['touch ' filename]);
 % Open file in append mode
 f = fopen(filename, 'a+');
@@ -11,8 +13,8 @@ f = fopen(filename, 'a+');
 baseSurvey = fileread('qualtricsBaseBlock.txt'); 
 
 % Load saved data to link to youtube
-data = STBData('SavedData');
-
+% data = STBData('SavedData');
+% data = survey;
 % Add header
 fprintf(f, '[[AdvancedFormat]]\n\n');
 for i = 1:length(data)
